# backend/src/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "MSTI Automation API",
    "version": "1.0.0",
    "description": "API documentation for MSTI Automation Backend"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "paths": {
    "/api/visualizations": {
      "get": {
        "summary": "Get all visualizations",
        "tags": ["Visualization"],
        "responses": {
          "200": {
            "description": "List of all visualizations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Visualization"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new visualization",
        "tags": ["Visualization"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "queryConfig": {
                    "type": "object"
                  },
                  "dataSourceId": {
                    "type": "string"
                  }
                },
                "required": ["name", "type"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visualization created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visualization"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/visualizations/{id}": {
      "get": {
        "summary": "Get visualization by ID",
        "tags": ["Visualization"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Visualization found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visualization"
                }
              }
            }
          },
          "404": {
            "description": "Visualization not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "put": {
        "summary": "Update visualization by ID",
        "tags": ["Visualization"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  },
                  "config": {
                    "type": "object"
                  },
                  "queryConfig": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Visualization updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Visualization"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "delete": {
        "summary": "Delete visualization by ID",
        "tags": ["Visualization"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Visualization deleted successfully"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/visualizations/flux-query": {
      "post": {
        "summary": "Execute raw Flux query",
        "tags": ["Visualization"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "dataSourceId": {
                    "type": "string"
                  },
                  "query": {
                    "type": "string"
                  },
                  "variables": {
                    "type": "object"
                  }
                },
                "required": ["dataSourceId", "query"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Query executed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "tags": ["User"],
        "responses": {
          "200": {
            "description": "List of all users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "tags": ["User"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password_hash": {
                    "type": "string"
                  },
                  "role": {
                    "type": "string"
                  }
                },
                "required": ["email", "password_hash", "role"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Visualization": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "config": {
            "type": "object"
          },
          "queryConfig": {
            "type": "object"
          },
          "dataSourceId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password_hash": {
            "type": "string"
          },
          "role": {
            "type": "string"
          }
        }
      }
    }
  }
}